optimized: formula cached
