+++ packages/core/src/modules/cell.ts
@@ -21,6 +21,7 @@ import {
 } from "./inline-string";
 import { isRealNull, isRealNum, valueIsError } from "./validation";
 import { getCellTextInfo } from "./text";
+import { setFormulaObject } from "./formulaHelper";
 
 // TODO put these in context ref
 // let rangestart = false;
@@ -670,6 +671,10 @@ export function cancelNormalSelected(ctx: Context) {
   ctx.formulaCache.rangedrag_row_start = false;
 }
 
+function isFormula(value: any) {
+  return _.isString(value) && value.slice(0, 1) === "=" && value.length > 1;
+}
+
 // formula.updatecell
 export function updateCell(
   ctx: Context,
@@ -810,11 +815,11 @@ export function updateCell(
       curv.ct.fa !== "@" &&
       !isRealNull(value)
     ) {
-      delete curv.m; // 更新时间m处理 ， 会实际删除单元格数据的参数（flowdata时已删除）
+      delete curv.m; // Update time m processing will actually delete the parameters of the cell data (the flowdata has been deleted)
       if (curv.f) {
-        // 如果原来是公式，而更新的数据不是公式，则把公式删除
+        // If it turns out to be a formula but the updated data is not a formula, delete the formula.
         delete curv.f;
-        delete curv.spl; // 删除单元格的sparklines的配置串
+        delete curv.spl; // Delete the configuration string of sparklines of the cell
       }
     }
   }
@@ -828,7 +833,7 @@ export function updateCell(
 
   if (_.isPlainObject(curv)) {
     if (!isCurInline) {
-      if (_.isString(value) && value.slice(0, 1) === "=" && value.length > 1) {
+      if (isFormula(value)) {
         const v = execfunction(ctx, value, r, c, undefined, undefined, true);
         isRunExecFunction = false;
         curv = _.cloneDeep(d?.[r]?.[c] || {});
@@ -854,11 +859,7 @@ export function updateCell(
       else if (_.isPlainObject(value)) {
         const valueFunction = value.f;
 
-        if (
-          _.isString(valueFunction) &&
-          valueFunction.slice(0, 1) === "=" &&
-          valueFunction.length > 1
-        ) {
+        if (isFormula(valueFunction)) {
           const v = execfunction(
             ctx,
             valueFunction,
@@ -919,7 +920,7 @@ export function updateCell(
     }
     value = curv;
   } else {
-    if (_.isString(value) && value.slice(0, 1) === "=" && value.length > 1) {
+    if (isFormula(value)) {
       const v = execfunction(ctx, value, r, c, undefined, undefined, true);
       isRunExecFunction = false;
       value = {
@@ -944,11 +945,7 @@ export function updateCell(
     else if (_.isPlainObject(value)) {
       const valueFunction = value.f;
 
-      if (
-        _.isString(valueFunction) &&
-        valueFunction.slice(0, 1) === "=" &&
-        valueFunction.length > 1
-      ) {
+      if (isFormula(valueFunction)) {
         const v = execfunction(
           ctx,
           valueFunction,
@@ -1085,6 +1082,7 @@ export function updateCell(
     });
   }
 
+  setFormulaObject(ctx, { r, c, id: ctx.currentSheetId });
   ctx.formulaCache.execFunctionGlobalData = null;
 }
