+++ packages/core/src/modules/toolbar.ts
@@ -41,6 +41,7 @@ import {
 import { showLinkCard } from "./hyperlink";
 import { cfSplitRange } from "./conditionalFormat";
 import { getCellTextInfo } from "./text";
+import { setFormulaObject } from "./formulaHelper";
 
 type ToolbarItemClickHandler = (
   ctx: Context,
@@ -901,6 +902,14 @@ export function autoSelectionFormula(
 
   if (!isfalse) {
     ctx.formulaCache.execFunctionExist.reverse();
+    ctx.formulaCache.execFunctionExist.forEach((formulaCell) => {
+      setFormulaObject(
+        ctx,
+        { r: formulaCell.r, c: formulaCell.c, id: ctx.currentSheetId },
+        flowdata
+      );
+    });
+
     // @ts-ignore
     execFunctionGroup(ctx, null, null, null, null, flowdata);
     ctx.formulaCache.execFunctionGlobalData = null;
